Add support for vector hashing
- Add support for hashing instances of VectorN classes, enabling their use as keys in unordered containers.
- Optimize TinyObjLoader::load by tracking duplicate vertices in an unordered map.
- Refactor MaterialXCore to remove unneeded dependencies.
